TourSchemSetup and overall reordering and modularizationg of the code
