+++ app/(root)/layout.tsx
@@ -1,4 +1,4 @@
-import Footer from "@/components/footer";
+import Footer from "@/components/shared/footer";
 import Header from "@/components/shared/header";
 
 export default function RootLayout({
