+++ components/shared/header/PageSection.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+
+const PageSection = () => {
+  return (
+    <div className="max-w-5xl mx-auto ">
+      <h2 className="text-4xl md:text-5xl font-bold  mb-6 font-serif">
+        Your Luxury Holiday Starts Here…
+      </h2>
+      <p className="text-lg md:text-xl  mb-2">
+        It’s time to elevate luxury travel out of the ordinary. Our experts have
+        travelled to the farthest reaches of the world, exploring everything
+        from the biggest cities to the smallest villages, from vast rainforests
+        to tiny desert islands. We use our inside knowledge to craft trips that
+        push the boundaries of your imagination. From enchanting honeymoons and
+        inspiring family holidays to monumental milestone adventures, our global
+        team of travel specialists are available to you 24/7 – ready to expertly
+        craft your{" "}
+        <a href="#" className="underline">
+          tailor-made luxury holiday
+        </a>{" "}
+        to your tastes.
+      </p>
+    </div>
+  );
+};
+
+export default PageSection;
