+++ components/shared/header/index.tsx
@@ -2,15 +2,12 @@
 import Link from "next/link";
 import { useState } from "react";
 import headerData from "@/data/header.json";
-import { getConfig } from "@/config/featureFlags";
 import { useRouter } from "next/navigation";
 import Image from "next/image";
 import Logo2 from "@/public/logos/logo_2.png";
 
 const Header = () => {
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
-  const isCountrySelectorVisible = getConfig("showCountrySelector");
-  const isCurrencySelectorVisible = getConfig("showCurrencySelector");
   const router = useRouter();
 
   const handleDestinationClick = (destinationLink: string) => {
@@ -234,44 +231,6 @@ const Header = () => {
                     />
                   </form>
                 </div>
-
-                {/* Mobile Selectors (if enabled) */}
-                {(isCountrySelectorVisible || isCurrencySelectorVisible) && (
-                  <div className="p-4 border-b border-gray-200 space-y-3">
-                    {isCountrySelectorVisible && (
-                      <div>
-                        <label className="block text-sm font-semibold text-gray-700 mb-2">
-                          {headerData.topBanner.countrySelector.label}
-                        </label>
-                        <select className="w-full px-3 py-2 border border-gray-300 rounded-md text-gray-800">
-                          {headerData.topBanner.countrySelector.options.map(
-                            (country) => (
-                              <option key={country} value={country}>
-                                {country}
-                              </option>
-                            )
-                          )}
-                        </select>
-                      </div>
-                    )}
-                    {isCurrencySelectorVisible && (
-                      <div>
-                        <label className="block text-sm font-semibold text-gray-700 mb-2">
-                          {headerData.topBanner.currencySelector.label}
-                        </label>
-                        <select className="w-full px-3 py-2 border border-gray-300 rounded-md text-gray-800">
-                          {headerData.topBanner.currencySelector.options.map(
-                            (currency) => (
-                              <option key={currency} value={currency}>
-                                {currency}
-                              </option>
-                            )
-                          )}
-                        </select>
-                      </div>
-                    )}
-                  </div>
-                )}
               </div>
 
               {/* Mobile CTA Button */}
